+++ src/google/protobuf/io/io_win32_unittest.cc
@@ -111,7 +111,7 @@ class IoWin32Test : public ::testing::Test {
 
 namespace {
 void StripTrailingSlashes(string* str) {
-  int i = str->size() - 1;
+  int i = static_cast<int>(str->size() - 1);
   for (; i >= 0 && ((*str)[i] == '/' || (*str)[i] == '\\'); --i) {}
   str->resize(i+1);
 }
